Refactor to better support usage as module
