+++ space2stats_api/src/space2stats/__init__.py
@@ -1,3 +1,6 @@
 """space2stats."""
 
+from .lib import StatsTable
+
+__all__ = ["StatsTable"]
 __version__ = "0.1.0"
