+++ space2stats_api/src/space2stats/api/errors.py
